fixing the bug in activity timer that was making it still cound on the exercises review page.
